Improve self-exclusion
The sed commands that were previously used would sometimes fail leaving
dependencies like "-x86_64-1gv" in the dep file. This problem would
manifest when a dependency's name started with the same string as the
current package. An example is running depfinder on the
totem-3.18.1-x86_64-1gv.txz
package, which had a dependency on
totem-pl-parser-3.10.7-x86_64-1gv.txz